job system created

+++ src/engine/lava_mesh.cpp
@@ -100,7 +100,7 @@ bool LavaMesh::loadAsGLTF(std::filesystem::path file_path){
 
         fastgltf::iterateAccessor<std::uint32_t>(gltf, indexaccessor,
           [&](std::uint32_t idx) {
-            indices.push_back(idx + initial_vtx);
+            indices.push_back(idx + (uint32_t)initial_vtx);
           });
       }
